Signs of life during cross section improvement
At the moment, there are cases where cross section improvement can take
many hours (we have a case of 5 hours that we're looking at right now).
When this sort of thing happens, it is easy for a user to think that the
job is stuck in an infinite loop if there is no other sign of life but
the CPU churning. This adds a little output every 5th refined job to let
folks know that things are proceeding.